pyproject.toml
[project]
name = "vaccfr-versioncheck"
version = "0.1.0"
description = "Automated version checker for Euroscope plugins"
requires-python = ">=3.10"
dependencies = [
    "PyGithub>=2.8.1",
    "packaging>=25",
]

[project.scripts]
version-check = "version_checker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["version_checker.py"]
